# crates/atuin-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atuin Sync Server API",
    "description": "HTTP API for the atuin sync server (crates/atuin-server). Schemas are derived from the Rust types in atuin-common, so they mirror their serde output exactly. All error responses use the ErrorResponse shape {\"reason\": string}. Every response carries an `Atuin-Version` header with the server's cargo version; the client uses it to refuse syncing across major versions. If the server is configured with a URL path prefix, it must be included in the client base URL.",
    "contact": {
      "name": "Ellie Huxtable",
      "email": "ellie@atuin.sh"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v0/record": {
      "get": {
        "tags": [
          "handlers::v0::record"
        ],
        "operationId": "record_status",
        "responses": {
          "200": {
            "description": "Max record index per host and tag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordStatus"
                }
              }
            }
          },
          "4XX": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "handlers::v0::record"
        ],
        "operationId": "post_records",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RecordEncrypted"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Records stored (empty body)"
          },
          "4XX": {
            "description": "Not authenticated, or a record was too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v0/record/next": {
      "get": {
        "tags": [
          "handlers::v0::record"
        ],
        "operationId": "next_records",
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "description": "Host ID (hyphenated lowercase UUID string)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Store tag, e.g. \"history\", \"kv\"",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Record index to start from; omitted means from the beginning",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Maximum number of records to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A batch of records (bare JSON array)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordEncrypted"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EncryptedData": {
        "type": "object",
        "required": [
          "data",
          "content_encryption_key"
        ],
        "properties": {
          "content_encryption_key": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "Host": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "RecordEncrypted": {
        "type": "object",
        "description": "A single record stored inside of our local database",
        "required": [
          "id",
          "idx",
          "host",
          "timestamp",
          "version",
          "tag",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EncryptedData"
          },
          "host": {
            "$ref": "#/components/schemas/Host"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "a unique ID"
          },
          "idx": {
            "type": "integer",
            "format": "uint64",
            "description": "The integer record ID. This is only unique per (host, tag).",
            "minimum": 0
          },
          "tag": {
            "type": "string",
            "description": "The type of data we are storing here. Eg, \"history\""
          },
          "timestamp": {
            "type": "integer",
            "format": "uint64",
            "description": "The creation time in nanoseconds since unix epoch",
            "minimum": 0
          },
          "version": {
            "type": "string",
            "description": "The version the data in the entry conforms to"
          }
        }
      },
      "RecordStatus": {
        "type": "object",
        "description": "An index representing the current state of the record stores\nThis can be both remote, or local, and compared in either direction",
        "required": [
          "hosts"
        ],
        "properties": {
          "hosts": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "uint64",
                "minimum": 0
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Session token presented as `Authorization: Token <session-token>`. This server accepts only the `Token` scheme (exactly one space separator); any other scheme is rejected with 400 {\"reason\":\"invalid authorization header encoding\"}. The full header value including the `Token ` prefix must be supplied."
      }
    }
  }
}